highscore working

+++ Assets/scripts/Save.cs
@@ -1,10 +1,9 @@
-using System.Collections;
-using System.Collections.Generic;
+
 using UnityEngine;
 
 [System.Serializable]
 public class Save 
 {
-   public float savedHighScore;
+   public int savedHighScore;
 }
